Create Beta-Binomial Empirical Bayes Bootstrap Correction.R
This code gives functions to perform a bootstrap correction of empirical Bayesian intervals for the beta-binomial model. The two main functions are mm.est(x,n) and mm.bootstrap(x,n), where x is a vector of counts in n trials. 

Two examples are provided at the bottom, using real and simulated data.